+++ resolvers/dns-resolver/Cargo.toml
@@ -15,14 +15,17 @@ edition = "2018"
 base64 = "0.10.1"
 chrono = "0.4"
 cueball = "0.3.0"
+custom_derive = "0.1.7"
 derive_more = "0.14.0"
 futures = "0.1"
+newtype_derive = "0.1.6"
 resolv-conf = "0.6.2"
 sha1 = "0.6.0"
 slog = "2.4.1"
 state_machine_future = "0.2.0"
-trust-dns-resolver = "0.11.1"
-trust-dns = "^0.16.1"
+thiserror = "1.0.15"
+trust-dns-client = "^0.19.4"
+trust-dns-proto = "0.19.4"
 tokio = "0.1.5"
 tokio-core = "0.1.17"
 unicode-normalization = "=0.1.5"
